requirements_dev.txt
-r requirements.txt

mypy>=1.4.0

homeassistant-stubs==2023.11.0
pytest-homeassistant-custom-component==0.13.74
